refactor: move settings_addons_list to a separate file
